# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification for Union Steward App
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /members/{id}/grievances:
    get:
      operationId: getMemberGrievances
      tags: [members]
      summary: Get grievances for a specific member
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of grievances for member
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Grievance"
  /grievances:
    get:
      operationId: listGrievances
      tags: [grievances]
      summary: List all grievances
      parameters:
        - in: query
          name: status
          schema:
            type: string
            enum: [open, pending_response, pending_hearing, resolved, withdrawn]
        - in: query
          name: step
          schema:
            type: integer
        - in: query
          name: memberId
          schema:
            type: integer
      responses:
        "200":
          description: List of grievances
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Grievance"
    post:
      operationId: createGrievance
      tags: [grievances]
      summary: File a new grievance
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateGrievanceBody"
      responses:
        "201":
          description: Filed grievance
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Grievance"
...
  /member-portal/profile:
    get:
      operationId: getApiMemberPortalProfile
      tags: [member-portal]
      summary: Get the current member's own profile
      responses:
        "200":
          description: Member profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MemberPortalProfile"
        "403":
          $ref: "#/components/responses/Forbidden"
    patch:
      operationId: patchApiMemberPortalProfile
      tags: [member-portal]
      summary: Update own notification preferences
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateMemberPortalProfileBody"
      responses:
        "200":
          description: Updated profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MemberPortalProfile"
  /member-portal/grievances:
    get:
      operationId: getApiMemberPortalGrievances
      tags: [member-portal]
      summary: List the current member's own grievances
      responses:
        "200":
          description: Member's grievances
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MemberPortalGrievance"
    post:
      operationId: postApiMemberPortalGrievances
      tags: [member-portal]
      summary: File a grievance as a member
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateMemberPortalGrievanceBody"
      responses:
        "201":
          description: Filed grievance
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MemberPortalGrievance"
...
  /member-portal/ai/chat:
    post:
      operationId: postApiMemberPortalAiChat
      tags: [member-portal]
      summary: Send messages to the CBA AI assistant (member role, SSE stream)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AiChatBody"
      responses:
        "200":
          description: SSE stream of assistant response
          content:
            text/event-stream: {}

  # ── PUSH NOTIFICATIONS ────────────────────────────────────────────────────────
...
  /members/{id}/discipline:
    get:
      operationId: getApiMembersIdDiscipline
      tags: [discipline]
      summary: List discipline records for a member (steward only)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Discipline records
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DisciplineRecord"
    post:
      operationId: postApiMembersIdDiscipline
      tags: [discipline]
      summary: Add a discipline record (steward only)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateDisciplineRecordBody"
      responses:
        "201":
          description: Created discipline record
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DisciplineRecord"
...
components:
  responses:
...
    Forbidden:
      description: Insufficient permissions
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
  schemas:
...
    Grievance:
      type: object
      properties:
        id:
          type: integer
        grievanceNumber:
          type: string
        memberId:
          type: integer
          nullable: true
        memberName:
          type: string
          nullable: true
        title:
          type: string
        description:
          type: string
          nullable: true
        contractArticle:
          type: string
          nullable: true
        step:
          type: integer
        status:
          type: string
          enum: [open, pending_response, pending_hearing, resolved, withdrawn]
        filedDate:
          type: string
          format: date
        dueDate:
          type: string
          format: date
          nullable: true
        resolvedDate:
          type: string
          format: date
          nullable: true
        resolution:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, grievanceNumber, title, step, status, filedDate, createdAt, updatedAt]
    CreateGrievanceBody:
      type: object
      properties:
        memberId:
          type: integer
          nullable: true
        title:
          type: string
        description:
          type: string
          nullable: true
        contractArticle:
          type: string
          nullable: true
        step:
          type: integer
          default: 1
        status:
          type: string
          enum: [open, pending_response, pending_hearing, resolved, withdrawn]
          default: open
        filedDate:
          type: string
          format: date
        dueDate:
          type: string
          format: date
          nullable: true
        notes:
          type: string
          nullable: true
      required: [title, filedDate]
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        code:
          type: string
      required: [error]
...
    MemberPortalProfile:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        employeeId:
          type: string
          nullable: true
        department:
          type: string
          nullable: true
        classification:
          type: string
          nullable: true
        phone:
          type: string
          nullable: true
        email:
          type: string
          nullable: true
        smsEnabled:
          type: boolean
        emailEnabled:
          type: boolean
        pushEnabled:
          type: boolean
      required: [id, name, smsEnabled, emailEnabled, pushEnabled]
    UpdateMemberPortalProfileBody:
      type: object
      properties:
        smsEnabled:
          type: boolean
        emailEnabled:
          type: boolean
        pushEnabled:
          type: boolean
    MemberPortalGrievance:
      type: object
      properties:
        id:
          type: integer
        grievanceNumber:
          type: string
        title:
          type: string
        status:
          type: string
          enum: [open, pending_response, pending_hearing, resolved, withdrawn]
        step:
          type: integer
        filedDate:
          type: string
          format: date
        dueDate:
          type: string
          format: date
          nullable: true
        resolvedDate:
          type: string
          format: date
          nullable: true
        resolution:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
      required: [id, grievanceNumber, title, status, step, filedDate, createdAt]
    CreateMemberPortalGrievanceBody:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
          nullable: true
        contractArticle:
          type: string
          nullable: true
        filedDate:
          type: string
          format: date
      required: [title, filedDate]
    AiChatMessage:
      type: object
      properties:
        role:
          type: string
          enum: [user, assistant]
        content:
          type: string
          minLength: 1
          maxLength: 10000
      required: [role, content]
    AiChatBody:
      type: object
      properties:
        messages:
          type: array
          items:
            $ref: "#/components/schemas/AiChatMessage"
          minItems: 1
          maxItems: 50
      required: [messages]

    # ── PUSH ──────────────────────────────────────────────────────────────────
...
    DisciplineRecord:
      type: object
      properties:
        id:
          type: integer
        memberId:
          type: integer
        disciplineType:
          type: string
          enum: [verbal_warning, written_warning, suspension_paid, suspension_unpaid, termination, other]
        incidentDate:
          type: string
          format: date
          nullable: true
        issuedDate:
          type: string
          format: date
        description:
          type: string
        responseFiled:
          type: boolean
        grievanceId:
          type: integer
          nullable: true
        createdBy:
          type: integer
          nullable: true
        createdAt:
          type: string
          format: date-time
      required: [id, memberId, disciplineType, issuedDate, description, responseFiled, createdAt]
    CreateDisciplineRecordBody:
      type: object
      properties:
        disciplineType:
          type: string
          enum: [verbal_warning, written_warning, suspension_paid, suspension_unpaid, termination, other]
        incidentDate:
          type: string
          format: date
          nullable: true
        issuedDate:
          type: string
          format: date
        description:
          type: string
        responseFiled:
          type: boolean
          default: false
        grievanceId:
          type: integer
          nullable: true
      required: [disciplineType, issuedDate, description]
...